+++ lib/quadrant_server.dart
@@ -4,11 +4,13 @@
 library quadrant_server;
 
 export 'src/quadrant_app.dart';
+export 'src/quadrant_router.dart';
 export 'src/middleware.dart';
 export 'src/request.dart';
 export 'src/response.dart';
 export 'src/route.dart';
 export 'src/websocket_context.dart';
+export 'src/websocket_group.dart';
 export 'src/websocket_route.dart';
 export 'src/middlewares/cors.dart';
 export 'src/middlewares/logger.dart';